References

GraphTensorNetworks.Coloring — Type

Coloring{K,CT<:EinTypes} <: GraphProblem
Coloring{K}(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

K-Coloring problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.ConfigEnumerator — Type

ConfigEnumerator{N,S,C}

Set algebra for enumerating configurations, where N is the length of configurations, C is the size of storage in unit of UInt64, S is the bit width to store a single element in a configuration, i.e. log2(# of flavors), for bitstrings, it is 1`.

julia> a = ConfigEnumerator([StaticBitVector([1,1,1,0,0]), StaticBitVector([1,0,0,0,1])])
{11100, 10001}

julia> b = ConfigEnumerator([StaticBitVector([0,0,0,0,0]), StaticBitVector([1,0,1,0,1])])
{00000, 10101}

julia> a + b
{11100, 10001, 00000, 10101}

GraphTensorNetworks.ConfigSampler — Type

ConfigSampler{N,S,C}

The algebra for sampling one configuration, where N is the length of configurations, C is the size of storage in unit of UInt64, S is the bit width to store a single element in a configuration, i.e. log2(# of flavors), for bitstrings, it is 1`.

julia> ConfigSampler(StaticBitVector([1,1,1,0,0]))
ConfigSampler{5, 1, 1}(11100)

GraphTensorNetworks.ConfigsAll — Type

ConfigsAll <:AbstractProperty
ConfigsAll()

Find all valid configurations, e.g. for Independence problem, it is finding all independent sets.

The corresponding data type is ConfigEnumerator.

Weights do not take effect.

GraphTensorNetworks.ConfigsMax — Type

ConfigsMax{K, BOUNDED} <:AbstractProperty
ConfigsMax(K=1; bounded=true)

Find configurations with largest sizes, e.g. for Independence problem, it is finding all independent sets of sizes $\alpha(G), \alpha(G)-1, \ldots, \alpha(G)-K+1$.

The corresponding data type is CountingTropical{Float64,<:ConfigEnumerator} for K == 1 and TruncatedPoly{K,<:ConfigEnumerator} for K > 1.

Weighted graph problems is only supported for K == 1.

GraphTensorNetworks.CountingAll — Type

CountingAll <: AbstractProperty
CountingAll()

Counting the total number of sets. e.g. for Independence problem, it counts the independent sets.

The corresponding tensor element type is Base.Real.

The weights on graph does not have effect.

BLAS (GPU and CPU) and GPU are supported,

GraphTensorNetworks.CountingMax — Type

CountingMax{K} <: AbstractProperty
CountingMax(K=1)

Counting the number of sets with K largest size. e.g. for Independence problem, it counts independent sets of size $\alpha(G), \alpha(G)-1, \ldots, \alpha(G)-K+1$.

The corresponding tensor element type is CountingTropical for K == 1, and TruncatedPoly{K} for K > 1.

Weighted graph problems is only supported for K == 1.

GPU is supported,

GraphTensorNetworks.GraphPolynomial — Type

GraphPolynomial{METHOD} <: AbstractProperty
GraphPolynomial(; method=:finitefield, kwargs...)

Compute the graph polynomial, e.g. for Independence problem, it is the independence polynomial. The METHOD type parameter can be one of the following symbols

:finitefield, it uses finite field algebra to fit the polynomial.

The corresponding tensor element type is Mods.Mod,

It does not have round-off error,

GPU is supported,

It accepts keyword arguments maxorder (optional, e.g. the MIS size in the Independence problem).

:polynomial, the program uses polynomial numbers to solve the polynomial directly.

The corresponding tensor element type is Polynomials.Polynomial.

It might have small round-off error depending on the data type for storing the counting.

It has memory overhead that linear to the graph size.

:fft,

The corresponding tensor element type is Base.Complex.

It has (controllable) round-off error.

BLAS and GPU are supported.

It accepts keyword arguments maxorder (optional) and r, if r > 1, one has better precision for coefficients of large order, if r < 1, one has better precision for coefficients of small order.

Graph polynomials are not defined for weighted graph problems.

GraphTensorNetworks.Independence — Type

Independence{CT<:EinTypes,WT<:Union{UnWeighted, Vector}} <: GraphProblem
Independence(graph; weights=UnWeighted(), openvertices=(),
             optimizer=GreedyMethod(), simplifier=nothing)

The Independent set problem. In the constructor, weights are the weights of vertices. openvertices specifies labels for the output tensor. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

An independent set is defined in the monadic second order logic as

\[\exists x_i,\ldots,x_M\left[\bigwedge_{i\neq j} (x_i\neq x_j \wedge \neg \textbf{adj}(x_i, x_j))\right]\]

In tensor network representation, we define a vertex tensors on vertex i (labeled by $s_i$) as

\[W(x_i)_{s_i} = \left(\begin{matrix} 1 \\ x_i \end{matrix}\right)_{s_i}\]

and a bond tensor on edge (i, j) as

\[B_{s_i s_j} = \left(\begin{matrix} 1 & 1\\ 1 & 0 \end{matrix}\right)_{s_is_j}\]

GraphTensorNetworks.Matching — Type

Matching{CT<:EinTypes} <: GraphProblem
Matching(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

Vertex matching problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details. The matching polynomial adopts the first definition in wiki page

\[m_G(x) := \sum_{k\geq 0}m_kx^k,\]

where m_k is the number of k-edge matchings.

GraphTensorNetworks.MaxCut — Type

MaxCut{CT<:EinTypes,WT<:Union{UnWeighted, Vector}} <: GraphProblem
MaxCut(graph; weights=UnWeighted(), openvertices=(),
            optimizer=GreedyMethod(), simplifier=nothing)

Max cut problem (or spin glass problem). In the constructor, weights are the weights of edges. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.MaximalIndependence — Type

MaximalIndependence{CT<:EinTypes,WT<:Union{UnWeighted, Vector}} <: GraphProblem
MaximalIndependence(graph; weights=UnWeighted(), openvertices=(),
             optimizer=GreedyMethod(), simplifier=nothing)

Maximal independent set problem. In the constructor, weights are the weights of vertices. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.PaintShop — Type

PaintShop{CT<:EinTypes} <: GraphProblem
PaintShop(labels::AbstractVector; openvertices=(),
         optimizer=GreedyMethod(), simplifier=nothing)

The binary paint shop problem.

Example

One can encode the paint shop problem abaccb as the following

julia> symbols = collect("abaccb");

julia> pb = PaintShop(symbols);

julia> solve(pb, SizeMax())[]
3.0ₜ

julia> solve(pb, ConfigsMax())[].c.data
2-element Vector{StaticBitVector{5, 1}}:
 01101
 01101

In our definition, we find the maximum number of unchanged color in this sequence, i.e. (n-1) - (minimum number of color changes) In the output of maximum configurations, the two configurations are defined on 5 bonds i.e. pairs of (i, i+1), 0 means color changed, while 1 means color not changed. If we denote two "colors" as r and b, then the optimal painting is rbbbrr or brrrbb, both change the colors twice.

GraphTensorNetworks.SingleConfigMax — Type

SingleConfigMax{BOUNDED} <: AbstractProperty
SingleConfigMax(; bounded=false)

Finding single best solution, e.g. for Independence problem, it is one of the maximum independent sets.

The corresponding data type is CountingTropical{Float64,<:ConfigSampler} if BOUNDED is true, Tropical otherwise.

Weighted graph problems is supported.

GPU is supported,

GraphTensorNetworks.SizeMax — Type

SizeMax <: AbstractProperty
SizeMax()

The maximum independent set size.

The corresponding tensor element type is Tropical.

It is compatible with weighted graph problems.

BLAS (on CPU) and GPU are supported,

GraphTensorNetworks.StaticElementVector — Type

StaticElementVector{N,S,C}

N is the length of vector, C is the size of storage in unit of UInt64, S is the stride defined as the log2(# of flavors). When the number of flavors is 2, it is a StaticBitVector.

GraphTensorNetworks.TruncatedPoly — Type

TruncatedPoly{K,T,TO} <: Number

Polynomial truncated to largest K orders. T is the coefficients type and TO is the orders type.

julia> TruncatedPoly((1,2,3), 6)
x^4 + 2*x^5 + 3*x^6

GraphTensorNetworks.all_solutions — Method

all_solutions(problem)

Finding all solutions. e.g. when the problem is MaximalIndependence, it computes all maximal independent sets, or the maximal cliques of it complement.

GraphTensorNetworks.backward_tropical — Method

backward_tropical(mode, ixs, xs, iy, y, ymask, size_dict)

The backward rule for tropical einsum.

mode can be one of :all and :single,

ixs and xs are labels and tensor data for input tensors,

iy and y are labels and tensor data for the output tensor,

ymask is the boolean mask for gradients,

size_dict is a key-value map from tensor label to dimension size.

GraphTensorNetworks.best2_solutions — Method

best2_solutions(problem; all=true, usecuda=false)

Finding optimal and suboptimal solutions.

GraphTensorNetworks.best_solutions — Method

best_solutions(problem; all=false, usecuda=false)

Find optimal solutions with bounding.

When all is true, the program will use set for enumerate all possible solutions, otherwise, it will return one solution for each size.

usecuda can not be true if you want to use set to enumerate all possible solutions.

GraphTensorNetworks.bounding_contract — Method

bounding_contract(mode, code, xsa, ymask, xsb; size_info=nothing)

Contraction method with bounding.

mode is a AllConfigs{K} instance, where MIS-K+1 is the largest IS size that you care about.

xsa are input tensors for bounding, e.g. tropical tensors,

xsb are input tensors for computing, e.g. tensors elements are counting tropical with set algebra,

ymask is the initial gradient mask for the output tensor.

GraphTensorNetworks.graph_polynomial — Function

graph_polynomial(problem, method; usecuda=false, kwargs...)

Computing the graph polynomial for specific problem.

problem can be one of the following instances,

Independence for the independence polynomial,

MaximalIndependence for the maximal independence polynomial,

Matching for the matching polynomial,

method can be one of the following inputs,

Val(:finitefield), compute exactly with the finite field method. It consumes additional kwargs [max_iter, maxorder], where maxorder is maximum order of polynomial and max_iter is the maximum number of primes numbers to use in the finitefield algebra. max_iter can be determined automatically in most cases.

Val(:polynomial), compute directly with Polynomial number type,

Val(:fft), compute with the fast fourier transformation approach, fast but needs to tune the hyperparameter r. It Consumes additional kwargs [maxorder, r]. The larger r is, the more accurate the factors of high order terms, and the less accurate the factors of low order terms.

Val(:fitting), compute with the polynomial fitting approach, fast but inaccurate for large graphs.

GraphTensorNetworks.labels — Method

labels(code)

Return a vector of unique labels in an Einsum token.

GraphTensorNetworks.load_configs — Method

load_configs(filename; format=:binary, bitlength=nothing, nflavors=2)

Load configurations from file filename. The format is :binary or :text. If the format is :binary, the bitstring length bitlength must be specified, nflavors specifies the degree of freedom.

GraphTensorNetworks.mis_compactify! — Method

mis_compactify!(tropicaltensor)

Compactify tropical tensor for maximum independent set problem. It will eliminate some entries by setting them to zero, by the criteria that even these entries are removed, the MIS size is not changed.

GraphTensorNetworks.save_configs — Method

save_configs(filename, data::ConfigEnumerator; format=:binary)

Save configurations data to file filename. The format is :binary or :text.

GraphTensorNetworks.set_packing — Method

set_packing(sets; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

Set packing is a generalization of independent set problem to hypergraphs. Calling this function will return you an Independence instance. sets are a vector of vectors, each element being a vertex in the independent set problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

Example

julia> sets = [[1, 2, 5], [1, 3], [2, 4], [3, 6], [2, 3, 6]];  # each set is a vertex

julia> gp = set_packing(sets);

julia> res = best_solutions(gp; all=true)[]
(2, {10010, 00110, 01100})ₜ

GraphTensorNetworks.show_graph — Method

show_graph(locations, edges;
    colors=["black", "black", ...],
    texts=["1", "2", ...],
    format=SVG,
    bond_color="black",
    )

Plots vertices at locations with colors specified by colors and texts specified by texts. You will need a VSCode, Pluto notebook or Jupyter notebook to show the image. If you want to write this image to the disk without displaying it in a frontend, please try

julia> open("test.png", "w") do f
            viz_atoms(f, generate_sites(SquareLattice(), 5, 5))
       end

The format keyword argument can also be Compose.SVG or Compose.PDF.

GraphTensorNetworks.solutions — Method

solutions(problem, basetype; all, usecuda=false)

General routine to find solutions without bounding,

basetype can be a type with counting field,

CountingTropical{Float64,Float64} for finding optimal solutions,

Polynomial{Float64, :x} for enumerating all solutions,

Max2Poly{Float64,Float64} for optimal and suboptimal solutions.

When all is true, the program will use set for enumerate all possible solutions, otherwise, it will return one solution for each size.

usecuda can not be true if you want to use set to enumerate all possible solutions.

GraphTensorNetworks.solve — Method

solve(problem, property; usecuda=false, T=Float64)

Solving a certain property of a graph problem.

Positional Arguments

problem is the graph problem with tensor network information,

property is string specifying the task. Using the maximum independent set problem as an example, it can be one of

SizeMax for finding maximum configuration size,

CountingMax for counting configurations with top K sizes,

CountingAll for counting all configurations,

GraphPolynomial for evaluating the graph polynomial,

SingleConfigMax for finding one maximum configuration,

ConfigsMax for enumerating configurations with top K sizes,

ConfigsAll for enumerating all configurations,

Keyword arguments

usecuda is a switch to use CUDA (if possible), user need to call statement using CUDA before turning on this switch.

T is the "base" element type, sometimes can be used to reduce the memory cost.

GraphTensorNetworks.@bv_str — Macro

Constructing a static bit vector.

TropicalNumbers.Tropical — Type

Tropical{T} <: Number

Tropical number is a semiring algebra that maps

+ in regular algebra to max,

* in regular algebra to +,

1 in regular algebra to 0,

0 in regular algebra to -Inf (for integer content types, this is chosen as a mall integer).

We implemented fast tropical matrix multiplication in TropicalGEMM.

TropicalNumbers.CountingTropical — Type

CountingTropical{T,CT} <: Number

Counting tropical number type is also a semiring algebra. It is tropical algebra with one extra field for counting, it is introduced in arXiv:2008.06888.

Mods.Mod — Type

Mod{m}(v) creates a modular number in mod m with value mod(v,m).

Polynomials.Polynomial — Type

Polynomial{T, X}(coeffs::AbstractVector{T}, [var = :x])

Construct a polynomial from its coefficients coeffs, lowest order first, optionally in terms of the given variable var which may be a character, symbol, or a string.

If $p = a_n x^n + \ldots + a_2 x^2 + a_1 x + a_0$, we construct this through Polynomial([a_0, a_1, ..., a_n]).

The usual arithmetic operators are overloaded to work with polynomials as well as with combinations of polynomials and scalars. However, operations involving two polynomials of different variables causes an error except those involving a constant polynomial.

Note

Polynomial is not axis-aware, and it treats coeffs simply as a list of coefficients with the first index always corresponding to the constant term. In order to use the axis of coeffs as exponents, consider using a LaurentPolynomial or possibly a SparsePolynomial.

Examples

julia> using Polynomials

julia> Polynomial([1, 0, 3, 4])
Polynomial(1 + 3*x^2 + 4*x^3)

julia> Polynomial([1, 2, 3], :s)
Polynomial(1 + 2*s + 3*s^2)

julia> one(Polynomial)
Polynomial(1.0)