References

GraphTensorNetworks.Coloring — Type

Coloring{K,CT<:EinTypes} <: GraphProblem
Coloring{K}(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

K-Coloring problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.Independence — Type

Independence{CT<:EinTypes} <: GraphProblem
Independence(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

Independent set problem. openvertices specifies the output tensor. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.Matching — Type

Matching{CT<:EinTypes} <: GraphProblem
Matching(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

Vertex matching problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details. The matching polynomial adopts the first definition in wiki page: https://en.wikipedia.org/wiki/Matching_polynomial

\[m_G(x) := \sum_{k\geq 0}m_kx^k,\]

where m_k is the number of k-edge matchings.

GraphTensorNetworks.MaxCut — Type

MaxCut{CT<:EinTypes} <: GraphProblem
MaxCut(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

Max cut problem (or spin glass problem). optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.MaximalIndependence — Type

MaximalIndependence{CT<:EinTypes} <: GraphProblem
MaximalIndependence(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

Maximal independent set problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.PaintShop — Type

PaintShop{CT<:EinTypes} <: GraphProblem
PaintShop(labels::AbstractVector; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

The binary paint shop problem: http://m-hikari.com/ams/ams-2012/ams-93-96-2012/popovAMS93-96-2012-2.pdf.

Example

One can encode the paint shop problem abaccb as the following

julia> symbols = collect("abaccb");

julia> pb = PaintShop(symbols);

julia> solve(pb, "size max")[]
3.0ₜ

julia> solve(pb, "configs max")[].c.data
2-element Vector{StaticBitVector{5, 1}}:
 01101
 01101

In our definition, we find the maximum number of unchanged color in this sequence, i.e. (n-1) - (minimum number of color changes) In the output of maximum configurations, the two configurations are defined on 5 bonds i.e. pairs of (i, i+1), 0 means color changed, while 1 means color not changed. If we denote two "colors" as r and b, then the optimal painting is rbbbrr or brrrbb, both change the colors twice.

GraphTensorNetworks.StaticElementVector — Type

StaticElementVector{N,S,C}

N is the length of vector, C is the size of storage in unit of UInt64, S is the stride defined as the log2(# of flavors). When the number of flavors is 2, it is a StaticBitVector.

GraphTensorNetworks.all_solutions — Method

all_solutions(problem)

Finding all solutions. e.g. when the problem is MaximalIndependence, it computes all maximal independent sets, or the maximal cliques of it complement.

GraphTensorNetworks.backward_tropical — Method

backward_tropical(mode, ixs, xs, iy, y, ymask, size_dict)

The backward rule for tropical einsum.

mode can be one of :all and :single,

ixs and xs are labels and tensor data for input tensors,

iy and y are labels and tensor data for the output tensor,

ymask is the boolean mask for gradients,

size_dict is a key-value map from tensor label to dimension size.

GraphTensorNetworks.best2_solutions — Method

best2_solutions(problem; all=true, usecuda=false)

Finding optimal and suboptimal solutions.

GraphTensorNetworks.best_solutions — Method

best_solutions(problem; all=false, usecuda=false)

Find optimal solutions with bounding.

When all is true, the program will use set for enumerate all possible solutions, otherwise, it will return one solution for each size.

usecuda can not be true if you want to use set to enumerate all possible solutions.

GraphTensorNetworks.bounding_contract — Method

bounding_contract(mode, code, xsa, ymask, xsb; size_info=nothing)

Contraction method with bounding.

* `mode` is a `AllConfigs{K}` instance, where `MIS-K+1` is the largest IS size that you care about.
* `xsa` are input tensors for bounding, e.g. tropical tensors,
* `xsb` are input tensors for computing, e.g. tensors elements are counting tropical with set algebra,
* `ymask` is the initial gradient mask for the output tensor.

GraphTensorNetworks.graph_polynomial — Function

graph_polynomial(problem, method; usecuda=false, kwargs...)

Computing the graph polynomial for specific problem.

problem can be one of the following instances,

Independence for the independence polynomial,

MaximalIndependence for the maximal independence polynomial,

Matching for the matching polynomial,

method can be one of the following inputs,

Val(:finitefield), compute exactly with the finite field method. It consumes additional kwargs [max_iter, maxorder], where maxorder is maximum order of polynomial and max_iter is the maximum number of primes numbers to use in the finitefield algebra. max_iter can be determined automatically in most cases.

Val(:polynomial), compute directly with Polynomial number type,

Val(:fft), compute with the fast fourier transformation approach, fast but needs to tune the hyperparameter r. It Consumes additional kwargs [maxorder, r]. The larger r is, the more accurate the factors of high order terms, and the less accurate the factors of low order terms.

Val(:fitting), compute with the polynomial fitting approach, fast but inaccurate for large graphs.

GraphTensorNetworks.labels — Method

labels(code)

Return a vector of unique labels in an Einsum token.

GraphTensorNetworks.load_configs — Method

load_configs(filename; format=:binary, bitlength=nothing, nflavors=2)

Load configurations from file filename. The format is :binary or :text. If the format is :binary, the bitstring length bitlength must be specified, nflavors specifies the degree of freedom.

GraphTensorNetworks.mis_compactify! — Method

mis_compactify!(tropicaltensor)

Compactify tropical tensor for maximum independent set problem. It will eliminate some entries by setting them to zero, by the criteria that even these entries are removed, the MIS size is not changed.

GraphTensorNetworks.save_configs — Method

save_configs(filename, data::ConfigEnumerator; format=:binary)

Save configurations data to file filename. The format is :binary or :text.

GraphTensorNetworks.set_packing — Method

set_packing(sets; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

Set packing is a generalization of independent set problem to hypergraphs. Calling this function will return you an Independence instance. sets are a vector of vectors, each element being a vertex in the independent set problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

Example

julia> sets = [[1, 2, 5], [1, 3], [2, 4], [3, 6], [2, 3, 6]];  # each set is a vertex

julia> gp = set_packing(sets);

julia> res = best_solutions(gp; all=true)[]
(2, {10010, 00110, 01100})ₜ

GraphTensorNetworks.solutions — Method

solutions(problem, basetype; all, usecuda=false)

General routine to find solutions without bounding,

basetype can be a type with counting field,

CountingTropical{Float64,Float64} for finding optimal solutions,

Polynomial{Float64, :x} for enumerating all solutions,

Max2Poly{Float64,Float64} for optimal and suboptimal solutions.

When all is true, the program will use set for enumerate all possible solutions, otherwise, it will return one solution for each size.

usecuda can not be true if you want to use set to enumerate all possible solutions.

GraphTensorNetworks.solve — Method

solve(problem, task; usecuda=false)

problem is the graph problem with tensor network information,

task is string specifying the task. Using the maximum independent set problem as an example, it can be one of

"size max", the maximum independent set size,

"counting sum", total number of independent sets,

"counting max", the dengeneracy of maximum independent sets (MIS),

"counting max2", the dengeneracy of MIS and MIS-1,

"counting all", independence polynomial, the polynomial number approach,

"counting all (fft)", independence polynomial, the fourier transformation approach,

"counting all (finitefield)", independence polynomial, the finite field approach,

"config max", one of the maximum independent set,

"config max (bounded)", one of the maximum independent set, the bounded version,

"configs max", all MIS configurations,

"configs max2", all MIS configurations and MIS-1 configurations,

"configs all", all IS configurations,

"configs max (bounded)", all MIS configurations, the bounded approach (much faster),

"configs max2 (bounded)", all MIS and MIS-1 configurations, the bounded approach (much faster),

"configs max3 (bounded)", all MIS, MIS-1 and MIS-2 configurations, the bounded approach (much faster),

GraphTensorNetworks.@bv_str — Macro

Constructing a static bit vector.