References

Graph problems

GraphTensorNetworks.solve — Function

solve(problem, property; usecuda=false, T=Float64)

Solving a certain property of a graph problem.

Positional Arguments

problem is the graph problem with tensor network information,

property is string specifying the task. Using the maximum independent set problem as an example, it can be one of

SizeMax for finding maximum set size,

SizeMin for finding minimum set size,

CountingMax for counting configurations with largest-K sizes,

CountingMin for counting configurations with smallest-K sizes,

CountingAll for counting all configurations,

GraphPolynomial for evaluating the graph polynomial,

SingleConfigMax for finding one maximum configuration,

ConfigsMax for enumerating configurations with largest-K sizes,

ConfigsMin for enumerating configurations with smallest-K sizes,

ConfigsAll for enumerating all configurations,

Keyword arguments

usecuda is a switch to use CUDA (if possible), user need to call statement using CUDA before turning on this switch.

T is the "base" element type, sometimes can be used to reduce the memory cost.

GraphTensorNetworks.GraphProblem — Type

GraphProblem

The abstract base type of graph problems.

GraphTensorNetworks.IndependentSet — Type

IndependentSet{CT<:AbstractEinsum,WT<:Union{UnWeighted, Vector}} <: GraphProblem
IndependentSet(graph; weights=UnWeighted(), openvertices=(),
             optimizer=GreedyMethod(), simplifier=nothing)

The independent set problem in graph theory. In the constructor, weights are the weights of vertices. openvertices specifies labels for the output tensor. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.MaximalIS — Type

MaximalIS{CT<:AbstractEinsum,WT<:Union{UnWeighted, Vector}} <: GraphProblem
MaximalIS(graph; weights=UnWeighted(), openvertices=(),
             optimizer=GreedyMethod(), simplifier=nothing)

The maximal independent set problem. In the constructor, weights are the weights of vertices. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.Matching — Type

Matching{CT<:AbstractEinsum} <: GraphProblem
Matching(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

The Vertex matching problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.Coloring — Type

Coloring{K,CT<:AbstractEinsum} <: GraphProblem
Coloring{K}(graph; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

The Vertex Coloring problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.MaxCut — Type

MaxCut{CT<:AbstractEinsum,WT<:Union{UnWeighted, Vector}} <: GraphProblem
MaxCut(graph; weights=UnWeighted(), openvertices=(),
            optimizer=GreedyMethod(), simplifier=nothing)

The cutting problem (or spin glass problem). In the constructor, weights are the weights of edges. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

GraphTensorNetworks.PaintShop — Type

PaintShop{CT<:AbstractEinsum} <: GraphProblem
PaintShop(sequence::AbstractVector; openvertices=(),
         optimizer=GreedyMethod(), simplifier=nothing)

The binary paint shop problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

Example

One can encode the paint shop problem abaccb as the following

julia> syms = collect("abaccb");

julia> pb = PaintShop(syms);

julia> solve(pb, SizeMin())[]
2.0ₜ

julia> solve(pb, ConfigsMin())[].c.data
2-element Vector{StaticBitVector{3, 1}}:
 100
 011

In our definition, we find the maximum number of unchanged color in this sequence, i.e. (n-1) - (minimum number of color changes) In the output of maximum configurations, the two configurations are defined on 5 bonds i.e. pairs of (i, i+1), 0 means color changed, while 1 means color not changed. If we denote two "colors" as r and b, then the optimal painting is rbbbrr or brrrbb, both change the colors twice.

GraphTensorNetworks.set_packing — Function

set_packing(sets; openvertices=(), optimizer=GreedyMethod(), simplifier=nothing)

Set packing is a generalization of independent set problem to hypergraphs. Calling this function will return you an IndependentSet instance. sets are a vector of vectors, each element being a vertex in the independent set problem. optimizer and simplifier are for tensor network optimization, check optimize_code for details.

Example

julia> sets = [[1, 2, 5], [1, 3], [2, 4], [3, 6], [2, 3, 6]];  # each set is a vertex

julia> gp = set_packing(sets);

julia> res = best_solutions(gp; all=true)[]
(2, {10010, 00110, 01100})ₜ

Graph Problem Interfaces

To subtype GraphProblem, a new type must contain a code field to represent the (optimized) tensor network. Interfaces GraphTensorNetworks.generate_tensors, labels, flavors and get_weights are required. nflavor is optional.

GraphTensorNetworks.generate_tensors — Function

generate_tensors(func, problem::GraphProblem)

Generate a vector of tensors as the inputs of the tensor network contraction code problem.code. func is a function to customize the tensors. func(symbol) returns a vector of elements, the length of which is same as the number of flavors.

Example

The following code gives your the maximum independent set size

julia> using Graphs, GraphTensorNetworks

julia> gp = IndependentSet(smallgraph(:petersen));

julia> getixsv(gp.code)
25-element Vector{Vector{Int64}}:
 [1]
 [2]
 [3]
 [4]
 [5]
 [6]
 [7]
 [8]
 [9]
 [10]
 ⋮
 [3, 8]
 [4, 5]
 [4, 9]
 [5, 10]
 [6, 8]
 [6, 9]
 [7, 9]
 [7, 10]
 [8, 10]

julia> gp.code(GraphTensorNetworks.generate_tensors(Tropical(1.0), gp)...)
0-dimensional Array{TropicalF64, 0}:
4.0ₜ

GraphTensorNetworks.labels — Function

labels(problem::GraphProblem)

The labels of a graph problem is defined as the degrees of freedoms in the graph problem. e.g. for the maximum independent set problems, they are the indices of vertices: 1, 2, 3..., while for the max cut problem, they are the edges.

GraphTensorNetworks.flavors — Function

flavors(::Type{<:GraphProblem})

It returns a vector of integers as the flavors of a degree of freedom. Its size is the same as the degree of freedom on a single vertex/edge.

GraphTensorNetworks.get_weights — Function

get_weights(problem::GraphProblem, sym)

The weights for the degree of freedom specified by sym of the graph problem, where sym is a symbol. In graph polynomial, integer weights are the orders of x.

GraphTensorNetworks.nflavor — Function

nflavor(::Type{<:GraphProblem})

Bond size is equal to the number of flavors.

Graph Problem Utilities

GraphTensorNetworks.is_independent_set — Function

is_independent_set(g::SimpleGraph, config)

Return true if config (a vector of boolean numbers as the mask of vertices) is an independent set of graph g.

GraphTensorNetworks.mis_compactify! — Function

mis_compactify!(tropicaltensor)

Compactify tropical tensor for maximum independent set problem. It will eliminate some entries by setting them to zero, by the criteria that removing these entry does not change the MIS size of its parent graph (reference to be added).

GraphTensorNetworks.is_maximal_independent_set — Function

is_maximal_independent_set(g::SimpleGraph, config)

Return true if config (a vector of boolean numbers as the mask of vertices) is a maximal independent set of graph g.

GraphTensorNetworks.cut_size — Function

cut_size(g::SimpleGraph, config; weights=UnWeighted())

Compute the cut size from vertex config (an iterator).

GraphTensorNetworks.num_paint_shop_color_switch — Function

num_paint_shop_color_switch(sequence::AbstractVector, coloring)

Returns the number of color switches.

GraphTensorNetworks.paint_shop_coloring_from_config — Function

paint_shop_coloring_from_config(p::PaintShop, config)

Returns a valid painting from the paint shop configuration (given by the configuration solvers). The config is a sequence of 0 and 1, where 0 means painting the first appearence of a car in blue, 1 otherwise.

GraphTensorNetworks.is_good_vertex_coloring — Function

is_good_vertex_coloring(graph::SimpleGraph, config)

Returns true if the coloring specified by config is a valid one, i.e. does not violate the contraints of vertices of an edges having different colors.

GraphTensorNetworks.CNF — Type

CNF{T}
CNF(clauses)

Boolean expression in conjunctive normal form. clauses is a vector of CNFClause, if and only if all clauses are satisfied, this CNF is satisfied.

Example

julia> @bools x y z

julia> cnf = (x ∨ y) ∧ (¬y ∨ z)
(x ∨ y) ∧ (¬y ∨ z)

julia> satisfiable(cnf, Dict([:x=>true, :y=>false, :z=>true]))
true

julia> satisfiable(cnf, Dict([:x=>false, :y=>false, :z=>true]))
false

GraphTensorNetworks.CNFClause — Type

CNFClause{T}
CNFClause(vars)

A clause in CNF, its value is the logical or of vars, where vars is a vector of BoolVar.

GraphTensorNetworks.BoolVar — Type

BoolVar{T}
BoolVar(name, neg)

Boolean variable for constructing CNF clauses.

GraphTensorNetworks.satisfiable — Function

satisfiable(cnf::CNF, config::AbstractDict)

Returns true if an assignment of variables satisfies a CNF.

GraphTensorNetworks.@bools — Macro

@bools(syms::Symbol...)

Create some boolean variables of type BoolVar in current scope that can be used in create a CNF.

Example

julia> @bools x y z

julia> (x ∨ y) ∧ (¬y ∨ z)
(x ∨ y) ∧ (¬y ∨ z)

GraphTensorNetworks.:∨ — Function

∨(vars...)

Logical or that applied on BoolVar and CNFClause and returns a [CNFClause].

GraphTensorNetworks.:¬ — Function

¬(var::BoolVar)

Negation of a boolean variables of type BoolVar.

GraphTensorNetworks.:∧ — Function

∧(vars...)

Logical and applied on CNFClause and CNF and returns a new [CNF].

Properties

GraphTensorNetworks.SizeMax — Type

SizeMax <: AbstractProperty
SizeMax()

The maximum set size. e.g. the largest size of the IndependentSet problem is also know as the independence number.

The corresponding tensor element type is max-plus tropical number Tropical.

It is compatible with weighted graph problems.

BLAS (on CPU) and GPU are supported,

GraphTensorNetworks.SizeMin — Type

SizeMin <: AbstractProperty
SizeMin()

The maximum set size. e.g. the smallest size ofthe MaximalIS problem is also known as the independent domination number.

The corresponding tensor element type inverted max-plus tropical number Tropical, which is equivalent to the min-plus tropical number.

It is compatible with weighted graph problems.

BLAS (on CPU) and GPU are supported,

GraphTensorNetworks.CountingAll — Type

CountingAll <: AbstractProperty
CountingAll()

Counting the total number of sets. e.g. for the IndependentSet problem, it counts the independent sets.

The corresponding tensor element type is Base.Real.

The weights on graph does not have effect.

BLAS (GPU and CPU) and GPU are supported,

GraphTensorNetworks.CountingMax — Type

CountingMax{K} <: AbstractProperty
CountingMax(K=1)

Counting the number of sets with largest-K size. e.g. for IndependentSet problem, it counts independent sets of size $\alpha(G), \alpha(G)-1, \ldots, \alpha(G)-K+1$.

The corresponding tensor element type is CountingTropical for K == 1, and TruncatedPoly{K} for K > 1.

Weighted graph problems is only supported for K == 1.

GPU is supported,

GraphTensorNetworks.CountingMin — Type

CountingMin{K} <: AbstractProperty
CountingMin(K=1)

Counting the number of sets with smallest-K size.

The corresponding tensor element type is inverted CountingTropical for K == 1, and TruncatedPoly{K} for K > 1.

Weighted graph problems is only supported for K == 1.

GPU is supported,

GraphTensorNetworks.GraphPolynomial — Type

GraphPolynomial{METHOD} <: AbstractProperty
GraphPolynomial(; method=:finitefield, kwargs...)

Compute the graph polynomial, e.g. for IndependentSet problem, it is the independence polynomial. The METHOD type parameter can be one of the following symbols

Method Argument

:finitefield, uses finite field algebra to fit the polynomial.

The corresponding tensor element type is Mods.Mod,

It does not have round-off error,

GPU is supported,

It accepts keyword arguments maxorder (optional, e.g. the MIS size in the IndependentSet problem).

:polynomial, use polynomial numbers to solve the polynomial directly.

The corresponding tensor element type is Polynomials.Polynomial.

It might have small round-off error depending on the data type for storing the counting.

It has memory overhead that linear to the graph size.

:fft, use fast fourier transformation to fit the polynomial.

The corresponding tensor element type is Base.Complex.

It has (controllable) round-off error.

BLAS and GPU are supported.

It accepts keyword arguments maxorder (optional) and r, if r > 1, one has better precision for coefficients of large order, if r < 1, one has better precision for coefficients of small order.

Graph polynomials are not defined for weighted graph problems.

GraphTensorNetworks.SingleConfigMax — Type

SingleConfigMax{BOUNDED} <: AbstractProperty
SingleConfigMax(; bounded=false)

Finding single best solution, e.g. for IndependentSet problem, it is one of the maximum independent sets.

The corresponding data type is CountingTropical{Float64,<:ConfigSampler} if BOUNDED is false, Tropical otherwise.

Weighted graph problems is supported.

GPU is supported,

Keyword Arguments

bounded, if it is true, use bounding trick (or boolean gradients) to reduce the working memory to store intermediate configurations.

GraphTensorNetworks.SingleConfigMin — Type

SingleConfigMin{BOUNDED} <: AbstractProperty
SingleConfigMin(; bounded=false)

Finding single "worst" solution.

The corresponding data type is inverted CountingTropical{Float64,<:ConfigSampler} if BOUNDED is false, inverted Tropical otherwise.

Weighted graph problems is supported.

GPU is supported,

Keyword Arguments

bounded, if it is true, use bounding trick (or boolean gradients) to reduce the working memory to store intermediate configurations.

GraphTensorNetworks.ConfigsAll — Type

ConfigsAll{TREESTORAGE} <:AbstractProperty
ConfigsAll(; tree_storage=false)

Find all valid configurations, e.g. for IndependentSet problem, it is finding all independent sets.

The corresponding data type is ConfigEnumerator.

Weights do not take effect.

Keyword Arguments

tree_storage, if it is true, it uses more memory efficient tree-structure to store the configurations.

GraphTensorNetworks.ConfigsMax — Type

ConfigsMax{K, BOUNDED, TREESTORAGE} <:AbstractProperty
ConfigsMax(K=1; bounded=true, tree_storage=true)

Find configurations with largest-K sizes, e.g. for IndependentSet problem, it is finding all independent sets of sizes $\alpha(G), \alpha(G)-1, \ldots, \alpha(G)-K+1$.

The corresponding data type is CountingTropical{Float64,<:ConfigEnumerator} for K == 1 and TruncatedPoly{K,<:ConfigEnumerator} for K > 1.

Weighted graph problems is only supported for K == 1.

Keyword Arguments

bounded, if it is true, use bounding trick (or boolean gradients) to reduce the working memory to store intermediate configurations.

tree_storage, if it is true, it uses more memory efficient tree-structure to store the configurations.

GraphTensorNetworks.ConfigsMin — Type

ConfigsMin{K, BOUNDED, TREESTORAGE} <:AbstractProperty
ConfigsMin(K=1; bounded=true, tree_storage=false)

Find configurations with smallest-K sizes.

The corresponding data type is inverted CountingTropical{Float64,<:ConfigEnumerator} for K == 1 and inverted TruncatedPoly{K,<:ConfigEnumerator} for K > 1.

Weighted graph problems is only supported for K == 1.

Keyword Arguments

bounded, if it is true, use bounding trick (or boolean gradients) to reduce the working memory to store intermediate configurations.

tree_storage, if it is true, it uses more memory efficient tree-structure to store the configurations.

Element Algebras

GraphTensorNetworks.is_commutative_semiring — Function

is_commutative_semiring(a::T, b::T, c::T) where T

Check if elements a, b and c satisfied the commutative semiring requirements.

\[\begin{align*} (a \oplus b) \oplus c = a \oplus (b \oplus c) & \hspace{5em}\triangleright\text{commutative monoid $\oplus$ with identity $\mathbb{0}$}\\ a \oplus \mathbb{0} = \mathbb{0} \oplus a = a &\\ a \oplus b = b \oplus a &\\ &\\ (a \odot b) \odot c = a \odot (b \odot c) & \hspace{5em}\triangleright \text{commutative monoid $\odot$ with identity $\mathbb{1}$}\\ a \odot \mathbb{1} = \mathbb{1} \odot a = a &\\ a \odot b = b \odot a &\\ &\\ a \odot (b\oplus c) = a\odot b \oplus a\odot c & \hspace{5em}\triangleright \text{left and right distributive}\\ (a\oplus b) \odot c = a\odot c \oplus b\odot c &\\ &\\ a \odot \mathbb{0} = \mathbb{0} \odot a = \mathbb{0} \end{align*}\]

TropicalNumbers.Tropical — Type

Tropical{T} <: Number

Tropical number is a semiring algebra that maps

+ in regular algebra to max,

* in regular algebra to +,

1 in regular algebra to 0,

0 in regular algebra to -Inf (for integer content types, this is chosen as a mall integer).

We implemented fast tropical matrix multiplication in TropicalGEMM.

Example

julia> Tropical(1.0) + Tropical(3.0)
3.0ₜ

julia> Tropical(1.0) * Tropical(3.0)
4.0ₜ

julia> one(TropicalF64)
0.0ₜ

julia> zero(TropicalF64)
-Infₜ

TropicalNumbers.CountingTropical — Type

CountingTropical{T,CT} <: Number

Counting tropical number type is also a semiring algebra. It is tropical algebra with one extra field for counting, it is introduced in arXiv:2008.06888.

Example

julia> CountingTropical(1.0, 5.0) + CountingTropical(3.0, 2.0)
(3.0, 2.0)ₜ

julia> CountingTropical(1.0, 5.0) * CountingTropical(3.0, 2.0)
(4.0, 10.0)ₜ

julia> one(CountingTropicalF64)
(0.0, 1.0)ₜ

julia> zero(CountingTropicalF64)
(-Inf, 0.0)ₜ

Mods.Mod — Type

Mod{m}(v) creates a modular number in mod m with value mod(v,m).

Polynomials.Polynomial — Type

Polynomial{T, X}(coeffs::AbstractVector{T}, [var = :x])

Construct a polynomial from its coefficients coeffs, lowest order first, optionally in terms of the given variable var which may be a character, symbol, or a string.

If $p = a_n x^n + \ldots + a_2 x^2 + a_1 x + a_0$, we construct this through Polynomial([a_0, a_1, ..., a_n]).

The usual arithmetic operators are overloaded to work with polynomials as well as with combinations of polynomials and scalars. However, operations involving two polynomials of different variables causes an error except those involving a constant polynomial.

Note

Polynomial is not axis-aware, and it treats coeffs simply as a list of coefficients with the first index always corresponding to the constant term. In order to use the axis of coeffs as exponents, consider using a LaurentPolynomial or possibly a SparsePolynomial.

Examples

julia> using Polynomials

julia> Polynomial([1, 0, 3, 4])
Polynomial(1 + 3*x^2 + 4*x^3)

julia> Polynomial([1, 2, 3], :s)
Polynomial(1 + 2*s + 3*s^2)

julia> one(Polynomial)
Polynomial(1.0)

GraphTensorNetworks.TruncatedPoly — Type

TruncatedPoly{K,T,TO} <: Number
TruncatedPoly(coeffs::Tuple, maxorder)

Polynomial truncated to largest K orders. T is the coefficients type and TO is the orders type.

Example

julia> TruncatedPoly((1,2,3), 6)
x^4 + 2*x^5 + 3*x^6

julia> TruncatedPoly((1,2,3), 6) * TruncatedPoly((5,2,1), 3)
20*x^7 + 8*x^8 + 3*x^9

julia> TruncatedPoly((1,2,3), 6) + TruncatedPoly((5,2,1), 3)
x^4 + 2*x^5 + 3*x^6

GraphTensorNetworks.Max2Poly — Type

Max2Poly{T,TO} = TruncatedPoly{2,T,TO}
Max2Poly(a, b, maxorder)

A shorthand of TruncatedPoly{2}.

GraphTensorNetworks.ConfigEnumerator — Type

ConfigEnumerator{N,S,C} <: AbstractSetNumber

Set algebra for enumerating configurations, where N is the length of configurations, C is the size of storage in unit of UInt64, S is the bit width to store a single element in a configuration, i.e. log2(# of flavors), for bitstrings, it is 1`.

Example

julia> a = ConfigEnumerator([StaticBitVector([1,1,1,0,0]), StaticBitVector([1,0,0,0,1])])
{11100, 10001}

julia> b = ConfigEnumerator([StaticBitVector([0,0,0,0,0]), StaticBitVector([1,0,1,0,1])])
{00000, 10101}

julia> a + b
{11100, 10001, 00000, 10101}

julia> one(a)
{00000}

julia> zero(a)
{}

GraphTensorNetworks.TreeConfigEnumerator — Type

TreeConfigEnumerator{N,S,C} <: AbstractSetNumber

Configuration enumerator encoded in a tree, it is the most natural representation given by a sum-product network and is often more memory efficient than putting the configurations in a vector. N, S and C are type parameters from the StaticElementVector{N,S,C}.

Fields

tag is one of ZERO, LEAF, SUM, PROD.

data is the element stored in a LEAF node.

left and right are two operands of a SUM or PROD node.

Example

julia> s = TreeConfigEnumerator(bv"00111")
00111

julia> q = TreeConfigEnumerator(bv"10000")
10000

julia> x = s + q
+
├─ 00111
└─ 10000

julia> y = x * x
*
├─ +
│  ├─ 00111
│  └─ 10000
└─ +
   ├─ 00111
   └─ 10000

julia> collect(y)
4-element Vector{StaticBitVector{5, 1}}:
 00111
 10111
 10111
 10000

julia> zero(s)

julia> one(s)
00000

GraphTensorNetworks.ConfigSampler — Type

ConfigSampler{N,S,C} <: AbstractSetNumber
ConfigSampler(elements::StaticElementVector)

The algebra for sampling one configuration, where N is the length of configurations, C is the size of storage in unit of UInt64, S is the bit width to store a single element in a configuration, i.e. log2(# of flavors), for bitstrings, it is 1`.

Note

ConfigSampler is a probabilistic commutative semiring, adding two config samplers do not give you deterministic results.

Example

julia> ConfigSampler(StaticBitVector([1,1,1,0,0]))
ConfigSampler{5, 1, 1}(11100)

julia> ConfigSampler(StaticBitVector([1,1,1,0,0])) + ConfigSampler(StaticBitVector([1,0,1,0,0]))
ConfigSampler{5, 1, 1}(10100)

julia> ConfigSampler(StaticBitVector([1,1,1,0,0])) * ConfigSampler(StaticBitVector([0,0,0,0,1]))
ConfigSampler{5, 1, 1}(11101)

julia> one(ConfigSampler{5, 1, 1})
ConfigSampler{5, 1, 1}(00000)

julia> zero(ConfigSampler{5, 1, 1})
ConfigSampler{5, 1, 1}(11111)

GraphTensorNetworks.StaticBitVector — Type

StaticBitVector{N,C} = StaticElementVector{N,1,C}
StaticBitVector(x::AbstractVector)

Example

julia> sb = StaticBitVector([1,0,0,1,1])
10011

julia> sb[3]
0x0000000000000000

GraphTensorNetworks.StaticElementVector — Type

StaticElementVector{N,S,C}
StaticElementVector(nflavor::Int, x::AbstractVector)

N is the length of vector, C is the size of storage in unit of UInt64, S is the stride defined as the log2(# of flavors). When the number of flavors is 2, it is a StaticBitVector.

Example

julia> ev = StaticElementVector(3, [1,2,0,1,2])
12012

julia> ev[2]
0x0000000000000002

GraphTensorNetworks.save_configs — Function

save_configs(filename, data::ConfigEnumerator; format=:binary)

Save configurations data to file filename. The format is :binary or :text.

GraphTensorNetworks.load_configs — Function

load_configs(filename; format=:binary, bitlength=nothing, nflavors=2)

Load configurations from file filename. The format is :binary or :text. If the format is :binary, the bitstring length bitlength must be specified, nflavors specifies the degree of freedom.

GraphTensorNetworks.@bv_str — Macro

Constructing a static bit vector.

GraphTensorNetworks.onehotv — Function

onehotv(::Type{<:StaticElementVector}, i, v)
onehotv(::Type{<:StaticBitVector, i)

Returns a static element vector, with the value at location i being v (1 if not specified).

Tensor Network

OMEinsumContractionOrders.optimize_code — Function

optimize_code(eincode, size_dict, optimizer = GreedyMethod(), simplifier=nothing, permute=true)

Optimize the einsum contraction code and reduce the time/space complexity of tensor network contraction. Returns a NestedEinsum instance. Input arguments are

eincode is an einsum contraction code instance, one of DynamicEinCode, StaticEinCode or NestedEinsum.

size is a dictionary of "edge label=>edge size" that contains the size information, one can use uniformsize(eincode, 2) to create a uniform size.

optimizer is a CodeOptimizer instance, should be one of GreedyMethod, KaHyParBipartite, SABipartite or TreeSA. Check their docstrings for details.

simplifier is one of MergeVectors or MergeGreedy.

optimize the permutation if permute is true.

OMEinsum.getixsv — Function

getixsv(code)

Get labels of input tensors for EinCode, NestedEinsum and some other einsum like objects. Returns a vector of vector.

julia> getixsv(ein"(ij,jk),k->i")
3-element Vector{Vector{Char}}:
 ['i', 'j']
 ['j', 'k']
 ['k']

OMEinsum.getiyv — Function

getiy(code)

Get labels of the output tensor for EinCode, NestedEinsum and some other einsum like objects. Returns a vector.

julia> getiyv(ein"(ij,jk),k->i")
1-element Vector{Char}:
 'i': ASCII/Unicode U+0069 (category Ll: Letter, lowercase)

OMEinsum.timespace_complexity — Function

timespace_complexity(eincode, size_dict)

Returns the time and space complexity of the einsum contraction. The time complexity is defined as log2(number of element multiplication). The space complexity is defined as log2(size of the maximum intermediate tensor).

OMEinsum.timespacereadwrite_complexity — Function

timespacereadwrite_complexity(eincode, size_dict)

Returns the time, space and read-write complexity of the einsum contraction. The time complexity is defined as log2(number of element-wise multiplication). The space complexity is defined as log2(size of the maximum intermediate tensor). The read-write complexity is defined as log2(the number of read-write operations).

OMEinsum.@ein_str — Macro

ein"ij,jk -> ik"(A,B)

String macro interface which understands numpy.einsum's notation. Translates strings into StaticEinCode-structs that can be called to evaluate an einsum. To control evaluation order, use parentheses - instead of an EinCode, a NestedEinsum is returned which evaluates the expression according to parens. The valid character ranges for index-labels are a-z and α-ω.

example

julia> a, b, c = rand(10,10), rand(10,10), rand(10,1);

julia> ein"ij,jk,kl -> il"(a,b,c) ≈ ein"(ij,jk),kl -> il"(a,b,c) ≈ a * b * c
true

OMEinsumContractionOrders.GreedyMethod — Type

GreedyMethod{MT}
GreedyMethod(; method=MinSpaceOut(), nrepeat=10)

The fast but poor greedy optimizer. Input arguments are

method is MinSpaceDiff() or MinSpaceOut.

MinSpaceOut choose one of the contraction that produces a minimum output tensor size,

MinSpaceDiff choose one of the contraction that decrease the space most.

nrepeat is the number of repeatition, returns the best contraction order.

OMEinsumContractionOrders.TreeSA — Type

TreeSA{RT,IT,GM}
TreeSA(; sc_target=20, βs=collect(0.01:0.05:15), ntrials=10, niters=50,
    sc_weight=1.0, rw_weight=0.2, initializer=:greedy, greedy_config=GreedyMethod(; nrepeat=1))

Optimize the einsum contraction pattern using the simulated annealing on tensor expression tree.

sc_target is the target space complexity,

ntrials, βs and niters are annealing parameters, doing ntrials indepedent annealings, each has inverse tempteratures specified by βs, in each temperature, do niters updates of the tree.

sc_weight is the relative importance factor of space complexity in the loss compared with the time complexity.

rw_weight is the relative importance factor of memory read and write in the loss compared with the time complexity.

initializer specifies how to determine the initial configuration, it can be :greedy or :random. If it is using :greedy method to generate the initial configuration, it also uses two extra arguments greedy_method and greedy_nrepeat.

nslices is the number of sliced legs, default is 0.

References

Recursive Multi-Tensor Contraction for XEB Verification of Quantum Circuits

OMEinsumContractionOrders.SABipartite — Type

SABipartite{RT,BT}
SABipartite(; sc_target=25, ntrials=50, βs=0.1:0.2:15.0, niters=1000
    max_group_size=40, greedy_config=GreedyMethod(), initializer=:random)

Optimize the einsum code contraction order using the Simulated Annealing bipartition + Greedy approach. This program first recursively cuts the tensors into several groups using simulated annealing, with maximum group size specifed by max_group_size and maximum space complexity specified by sc_target, Then finds the contraction order inside each group with the greedy search algorithm. Other arguments are

size_dict, a dictionary that specifies leg dimensions,

sc_target is the target space complexity, defined as log2(number of elements in the largest tensor),

max_group_size is the maximum size that allowed to used greedy search,

βs is a list of inverse temperature 1/T,

niters is the number of iteration in each temperature,

ntrials is the number of repetition (with different random seeds),

greedy_config configures the greedy method,

initializer, the partition configuration initializer, one can choose :random or :greedy (slow but better).

References

Hyper-optimized tensor network contraction

OMEinsumContractionOrders.KaHyParBipartite — Type

KaHyParBipartite{RT,IT,GM}
KaHyParBipartite(; sc_target, imbalances=collect(0.0:0.005:0.8),
    max_group_size=40, greedy_config=GreedyMethod())

Optimize the einsum code contraction order using the KaHyPar + Greedy approach. This program first recursively cuts the tensors into several groups using KaHyPar, with maximum group size specifed by max_group_size and maximum space complexity specified by sc_target, Then finds the contraction order inside each group with the greedy search algorithm. Other arguments are

sc_target is the target space complexity, defined as log2(number of elements in the largest tensor),

imbalances is a KaHyPar parameter that controls the group sizes in hierarchical bipartition,

max_group_size is the maximum size that allowed to used greedy search,

greedy_config is a greedy optimizer.

References

Hyper-optimized tensor network contraction

Simulating the Sycamore quantum supremacy circuits

OMEinsumContractionOrders.MergeVectors — Type

MergeVectors <: CodeSimplifier
MergeVectors()

Contraction code simplifier (in order to reduce the time of calling optimizers) that merges vectors to closest tensors.

OMEinsumContractionOrders.MergeGreedy — Type

MergeGreedy <: CodeSimplifier
MergeGreedy(; threshhold=-1e-12)

Contraction code simplifier (in order to reduce the time of calling optimizers) that merges tensors greedily if the space complexity of merged tensors is reduced (difference smaller than the threshhold).

Others

Graph

GraphTensorNetworks.show_graph — Function

show_graph(locations, edges;
    vertex_colors=["black", "black", ...],
    edge_colors=["black", "black", ...],
    texts=["1", "2", ...],
    format=SVG,
    edge_color="black",
    kwargs...
    )
show_graph(graph::SimpleGraph; locs=spring_layout(graph), kwargs...)

Plots vertices at locations with vertex colors specified by vertex_colors and texts specified by texts. You will need a VSCode, Pluto notebook or Jupyter notebook to show the image. If you want to write this image to the disk without displaying it in a frontend, please try

julia> open("test.png", "w") do f
            viz_atoms(f, generate_sites(SquareLattice(), 5, 5))
       end

The format keyword argument can be Compose.SVG or Compose.PDF.

Other keyword arguments

line, vertex and text

scale::Float64 = 1.0

pad::Float64 = 1.5

vertex

vertex_text_color::String = "black"

vertex_stroke_color = "black"

vertex_fill_color = "white"

edge

edge_color::String = "black"

image size in cm

image_size::Float64 = 12

GraphTensorNetworks.spring_layout — Function

Spring layout for graph plotting, returns a vector of vertex locations.

Note

This function is copied from GraphPlot.jl, where you can find more information about his function.

GraphTensorNetworks.diagonal_coupled_graph — Function

diagonal_coupled_graph(mask::AbstractMatrix{Bool})

Create a masked diagonal coupled square lattice graph from a specified mask.

GraphTensorNetworks.square_lattice_graph — Function

square_lattice_graph(mask::AbstractMatrix{Bool})

Create a masked square lattice graph.

GraphTensorNetworks.unit_disk_graph — Function

unit_disk_graph(locs::AbstractVector, unit::Real)

Create a unit disk graph with locations specified by locs and unit distance unit.

GraphTensorNetworks.line_graph — Function

line_graph(g::SimpleGraph)

Returns the line graph of g. The line graph is generated by mapping an edge to a vertex and two edges sharing a common vertex will be connected.

GraphTensorNetworks.random_diagonal_coupled_graph — Function

random_diagonal_coupled_graph(m::Int, n::Int, ρ::Real)

Create a random masked diagonal coupled square lattice graph, with number of vertices fixed to $\lfloor mn\rho \rceil$.

GraphTensorNetworks.random_square_lattice_graph — Function

random_square_lattice_graph(m::Int, n::Int, ρ::Real)

Create a random masked square lattice graph, with number of vertices fixed to $\lfloor mn\rho \rceil$.

One can also use random_regular_graph and smallgraph in Graphs to build special graphs.

Shortcuts

GraphTensorNetworks.max_size — Function

max_size(problem; usecuda=false)

Returns the maximum size of the graph problem. A shorthand of solve(problem, SizeMax(); usecuda=false).

GraphTensorNetworks.max_size_count — Function

max_size_count(problem; usecuda=false)

Returns the maximum size and the counting of the graph problem. It is a shorthand of solve(problem, CountingMax(); usecuda=false).